+++ ntex-error/src/message.rs
@@ -1,4 +1,4 @@
-use std::{error, fmt, fmt::Write, rc::Rc};
+use std::{error::Error as StdError, fmt, fmt::Write, rc::Rc};
 
 use ntex_bytes::ByteString;
 
@@ -33,13 +33,13 @@ impl fmt::Write for Wrt<'_> {
     }
 }
 
-pub fn fmt_err_string(e: &dyn error::Error) -> String {
+pub fn fmt_err_string(e: &dyn StdError) -> String {
     let mut buf = String::new();
     _ = fmt_err(&mut buf, e);
     buf
 }
 
-pub fn fmt_err(f: &mut dyn fmt::Write, e: &dyn error::Error) -> fmt::Result {
+pub fn fmt_err(f: &mut dyn fmt::Write, e: &dyn StdError) -> fmt::Result {
     let mut wrt = Wrt::new(f);
     let mut current = Some(e);
     while let Some(std_err) = current {
@@ -55,7 +55,7 @@ pub fn fmt_err(f: &mut dyn fmt::Write, e: &dyn error::Error) -> fmt::Result {
 /// Formats a full diagnostic view of an error for logging and tracing.
 pub fn fmt_diag_string<'a, T>(e: &'a T) -> String
 where
-    T: ErrorDiagnostic + AsError,
+    T: AsError,
     ResultType: From<&'a T::Target>,
 {
     let mut buf = String::new();
@@ -69,7 +69,7 @@ where
 /// and a backtrace when available.
 pub fn fmt_diag<'a, T>(f: &mut dyn fmt::Write, container: &'a T) -> fmt::Result
 where
-    T: ErrorDiagnostic + AsError,
+    T: AsError,
     ResultType: From<&'a T::Target>,
 {
     fmt_diag_typ(f, Some(ResultType::from(container.as_diag())), container)
@@ -79,10 +79,12 @@ where
 ///
 /// For `ServiceError` types, this includes debug representations of all nested errors,
 /// and a backtrace when available.
-pub fn fmt_diag_typ<T>(f: &mut dyn fmt::Write, typ: Option<ResultType>, e: &T) -> fmt::Result
+pub fn fmt_diag_typ<T>(f: &mut dyn fmt::Write, typ: Option<ResultType>, err: &T) -> fmt::Result
 where
-    T: ErrorDiagnostic,
+    T: AsError,
 {
+    let e = err.as_diag();
+
     writeln!(f, "err: {e}")?;
     if let Some(ref tp) = typ {
         writeln!(f, "type: {}", tp.as_str())?;
@@ -132,14 +134,14 @@ pub struct ErrorMessage(ByteString);
 #[derive(Clone)]
 pub struct ErrorMessageChained {
     msg: ByteString,
-    source: Option<Rc<dyn error::Error>>,
+    source: Option<Rc<dyn StdError>>,
 }
 
 impl ErrorMessageChained {
     pub fn new<M, E>(ctx: M, source: E) -> Self
     where
         M: Into<ErrorMessage>,
-        E: error::Error + 'static,
+        E: StdError + 'static,
     {
         ErrorMessageChained {
             msg: ctx.into().into_string(),
@@ -189,7 +191,7 @@ impl ErrorMessage {
         self.0
     }
 
-    pub fn with_source<E: error::Error + 'static>(self, source: E) -> ErrorMessageChained {
+    pub fn with_source<E: StdError + 'static>(self, source: E) -> ErrorMessageChained {
         ErrorMessageChained::new(self, source)
     }
 }
@@ -245,8 +247,8 @@ impl<M: Into<ErrorMessage>> From<M> for ErrorMessageChained {
     }
 }
 
-impl error::Error for ErrorMessageChained {
-    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
+impl StdError for ErrorMessageChained {
+    fn source(&self) -> Option<&(dyn StdError + 'static)> {
         self.source.as_ref().map(AsRef::as_ref)
     }
 }