+++ ntex/src/web/app_service.rs
@@ -1,6 +1,6 @@
 use std::{cell::RefCell, marker, mem, rc::Rc};
 
-use crate::error::ErrorInfo;
+use crate::error::Failure;
 use crate::http::{Request, Response};
 use crate::router::{Path, ResourceDef, ResourceId, Router};
 use crate::service::cfg::{Cfg, Configuration};
@@ -23,7 +23,7 @@ type Guards = Vec<Box<dyn Guard>>;
 pub struct AppFactory<St, M, F>
 where
     St: AppState,
-    F: ServiceFactory<St, WebRequest, Res = WebRequest, Error = St::Error, InitError = ErrorInfo>,
+    F: ServiceFactory<St, WebRequest, Res = WebRequest, Error = St::Error, InitError = Failure>,
 {
     middleware: M,
     filter: ServiceChainFactory<F, St, WebRequest>,
@@ -37,7 +37,7 @@ where
     St: AppState,
     M: Middleware<AppRouter<St, F::Service>, St> + 'static,
     M::Service: Service<St, WebRequest, Res = WebResponse, Error = St::Error>,
-    F: ServiceFactory<St, WebRequest, Res = WebRequest, Error = St::Error, InitError = ErrorInfo>,
+    F: ServiceFactory<St, WebRequest, Res = WebRequest, Error = St::Error, InitError = Failure>,
 {
     pub(super) fn new(
         middleware: M,
@@ -109,13 +109,13 @@ where
     St: AppState,
     M: Middleware<AppRouter<St, F::Service>, St> + 'static,
     M::Service: Service<St, WebRequest, Res = WebResponse, Error = St::Error>,
-    F: ServiceFactory<St, WebRequest, Res = WebRequest, Error = St::Error, InitError = ErrorInfo>,
+    F: ServiceFactory<St, WebRequest, Res = WebRequest, Error = St::Error, InitError = Failure>,
 {
     type Res = WebResponse;
     type Error = St::Error;
 
     type Service = AppService<M::Service, St>;
-    type InitError = ErrorInfo;
+    type InitError = Failure;
 
     async fn create(&self, st: &St) -> Result<Self::Service, Self::InitError> {
         let filter = self.filter.create(st).await?;